Add vhost-user-sound crate
