+++ crates/sound/CHANGELOG.md
@@ -1,0 +1,3 @@
+# Upcoming Release
+
+- First initial daemon implementation.
