+++ crates/sound/src/audio_backends/pipewire.rs
@@ -1,17 +1,18 @@
+// Pipewire backend device
 // SPDX-License-Identifier: Apache-2.0 or BSD-3-Clause
 
 use super::AudioBackend;
 use crate::{Error, Result, SoundRequest};
 
-pub struct NullBackend {}
+pub struct PwBackend {}
 
-impl NullBackend {
+impl PwBackend {
     pub fn new() -> Self {
         Self {}
     }
 }
 
-impl AudioBackend for NullBackend {
+impl AudioBackend for PwBackend {
     fn write(&self, _req: &SoundRequest) -> Result<()> {
         Ok(())
     }
